+++ src/main/java/com/strictmanager/travelbudget/domain/plan/service/PlanService.java
@@ -1,16 +1,39 @@
 package com.strictmanager.travelbudget.domain.plan.service;
 
-import com.strictmanager.travelbudget.domain.budget.Budget;
+import com.strictmanager.travelbudget.domain.plan.TripMember;
 import com.strictmanager.travelbudget.domain.plan.TripPlan;
+import com.strictmanager.travelbudget.domain.plan.TripPlan.YnFlag;
+import com.strictmanager.travelbudget.infra.persistence.jpa.TripMemberRepository;
+import com.strictmanager.travelbudget.infra.persistence.jpa.TripPlanRepository;
+import java.util.Comparator;
 import java.util.List;
+import java.util.stream.Collectors;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 
-public interface PlanService {
+@Service
+@RequiredArgsConstructor
+public class PlanService {
 
-    List<TripPlan> getPlans(Long userId);
+    private final TripPlanRepository tripPlanRepository;
+    private final TripMemberRepository tripMemberRepository;
 
-    Budget createBudget(Budget budget);
+    public List<TripPlan> getPlans(Long userId) {
+        return tripMemberRepository.findByUser_Id(userId)
+            .stream()
+            .map(TripMember::getTripPlan)
+            .filter(tripPlan -> tripPlan.getIsDelete().equals(YnFlag.N))
+            .sorted(Comparator.comparing(TripPlan::getStartDate).reversed())
+            .collect(Collectors.toList());
+    }
 
-    TripPlan createPlan(TripPlan tripPlan);
 
-    TripPlan createNewPlan(Long userId);
+    public TripPlan createPlan(TripPlan tripPlan) {
+        return tripPlanRepository.save(tripPlan);
+    }
+
+    public TripMember createTripMember(TripMember tripMember) {
+        return tripMemberRepository.save(tripMember);
+    }
+
 }
