+++ src/main/java/com/strictmanager/travelbudget/web/PlanController.java
@@ -4,15 +4,16 @@ package com.strictmanager.travelbudget.web;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.strictmanager.travelbudget.domain.budget.Budget;
+import com.strictmanager.travelbudget.domain.budget.BudgetService;
+import com.strictmanager.travelbudget.domain.plan.TripMember;
+import com.strictmanager.travelbudget.domain.plan.TripMember.Authority;
 import com.strictmanager.travelbudget.domain.plan.TripPlan;
 import com.strictmanager.travelbudget.domain.plan.service.PlanService;
 import com.strictmanager.travelbudget.domain.user.User;
 import java.net.URI;
 import java.time.LocalDate;
 import java.util.List;
-import java.util.Objects;
 import java.util.Optional;
-import java.util.OptionalLong;
 import javax.servlet.http.HttpServletRequest;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -26,12 +27,11 @@ import org.springframework.web.bind.annotation.RequestBody;
 
 @Slf4j
 @ApiController
-//@DevController // TODO: 개발 단계에서 사용 2020-07-25 (kiyeon_kim1)
 @RequiredArgsConstructor
 public class PlanController {
 
     private final PlanService planService;
-    private final static Long DEV_USER_ID = 1L; // TODO: 개발단계에서 사 2020-07-25 (kiyeon_kim1)
+    private final BudgetService budgetService;
 
     @GetMapping("/plans")
     public ResponseEntity<List<TripPlan>> getUserPlans(@AuthenticationPrincipal User user) {
@@ -47,12 +47,12 @@ public class PlanController {
         Optional<Long> sharedBudgetOpt = Optional.ofNullable(param.getSharedBudget());
 
         Budget budget = sharedBudgetOpt
-            .map(amount -> planService.createBudget(Budget.builder()
+            .map(amount -> budgetService.createBudget(Budget.builder()
                 .amount(amount)
                 .build()))
             .orElse(null);
 
-        planService.createPlan(TripPlan.builder()
+        TripPlan tripPlan = planService.createPlan(TripPlan.builder()
             .name(param.name)
             .startDate(param.getStartDate())
             .endDate(param.getEndDate())
@@ -60,7 +60,12 @@ public class PlanController {
             .userId(user.getId())
             .build());
 
-        // TODO: 영속성 컨텍스트 관련 수정 필요!! 2020-07-26 (kiyeon_kim1)
+        planService.createTripMember(TripMember.builder()
+            .authority(Authority.OWNER)
+            .tripPlan(tripPlan)
+            .budget(budget)
+            .user(user)
+            .build());
 
         return ResponseEntity
             .created(URI.create(httpServletRequest.getRequestURI()))
@@ -76,8 +81,6 @@ public class PlanController {
         private final LocalDate endDate;
 
         private final Long sharedBudget;
-        //        private final Long personalBudget;
-//        private final boolean isPublic;
 
         @JsonCreator
         public PlanCreateRequest(
@@ -85,16 +88,11 @@ public class PlanController {
             @JsonProperty(value = "startDate", required = true) LocalDate startDate,
             @JsonProperty(value = "endDate", required = true) LocalDate endDate,
             @JsonProperty(value = "sharedBudget", required = false) Long sharedBudget
-//            @JsonProperty(value = "personalBudget", required = false) Long personalBudget,
-//            @JsonProperty(value = "isPublic", required = true, defaultValue = "false") boolean isPublic
         ) {
             this.name = name;
             this.startDate = startDate;
             this.endDate = endDate;
             this.sharedBudget = sharedBudget;
-            //Objects.requireNonNullElse(sharedBudget, 0L);
-//            this.personalBudget = personalBudget;
-//            this.isPublic = isPublic;
         }
     }
